Fixed font rendering, added image test window
Fonts now render cleanly regardless of MG sampling being used.  Previously the fonts were blurry as the renderer was applying a 0.5f offset to the MG effect's projection matrix.  This was required due to some oddities in the OGL renderer but would cause blurryness in the DX renderer.  This inconsistency was fixed in MG 3.8 and now no offset is required.

This also added a window to test images in ImGuiHandler.